+++ packages/cinterop/src/jvm/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -741,12 +741,12 @@ actual object RealmInterop {
     actual fun realm_app_sync_client_get_default_file_path_for_realm(
         app: RealmAppPointer,
         syncConfig: RealmSyncConfigurationPointer,
-        name: String?
+        overriddenName: String?
     ): String {
         return realmc.realm_app_sync_client_get_default_file_path_for_realm(
             app.cptr(),
             syncConfig.cptr(),
-            name
+            overriddenName
         )
     }